improved singleton bean instantiation
The way I was doing it was wrong, I mean, it was working but, I took the
wrong approach of using nested for loops to detect and work on inner
dependencies. However this approach I took is not fine since I could
only go through N levels of inner dependencies. So now I'm using a
recursive approach, which doesn't matter how many dependencies a class
has, it would go through all of them.