fixed ambiguity resolution when multiple classes implements the same interface
While prior to this oblivion was working fine to inject classes directly
referenced in the constructor, there was a problem when we passed an
interface instead. We can't inject an interface, and if there was
multiple implementations of that interface, there was no way to know
which one to work on.

Solved this by introducing a new annotation '@OblivionPrimary'. Later I
intend to expand this and add a more robust Qualifier implementation,
similar to Spring.